requirements.txt
Flask[async]==2.3.3
python-dotenv==1.0.1
requests==2.31.0
Flask-Login==0.6.3
authlib==1.3.0
Flask-Session==0.5.0
cachelib==0.10.2
mongoengine==0.27.0
pymongo==4.5.0
cursor-agent
ollama
packaging 